W172c: br_root 001F64C0 260/288 nd144 -> 276/288 MATCH (transferred the matching sibling idioms: one-case switch for the positive out-of-line state-1 body, signed loop bounds, staged entry-address calculation, typed pointer-base-first addition). File 56 -> 57 MATCH.
ADJUDICATION: the agent left three unfinished experiments and I reverted two by marker-bounded byte splice. 001F6A60 had become 716/704 -- a NEW over-window addition, which is an absolute rule violation -- restored to 700/704 nd183 even though its nd had improved to 115; the agent's own note says retail needs physical case bodies 0,2,1,default, so that line is worth resuming properly rather than leaving over-window. 001F1240 had drifted 360/368 -> 316/368: nd improved 235 -> 226 but the object moved 44 bytes AWAY from its window, which is the wrong direction for an undersized function; restored from the pre-wave revision. 001F0C40 (284/304 nd161 vs 288/304 nd159) is within noise and its retail flag-clear correction is sound, so it stays.

Kept the byte-neutral SflResourceManager* cast at the sflResInit call site -- the concurrently-typed header otherwise breaks br_root's compilation. File now 57 MATCH, one over-window (001F6D20, pre-existing), zero COMPILE_ERROR.

+++ src/itfMesManager.c
@@ -188,7 +188,7 @@ void FUN_003a5540(int p1,int p2,u64 p3,int p4,int p5,int p6,int p7);
 extern void FUN_003b0ce0_typed(u64 param_1,u64 param_2);
 void FUN_003a5540(int p1,int p2,u64 p3,int p4,int p5,int p6,int p7);
 u64  FUN_003a5570(u32 param_1,u64 param_2,u64 param_3,u32 param_4,  u64 param_5,long param_6,u64 param_7,u32 param_8,u64 param_9  );
-u64  FUN_003a56f0(u32 param_1,u64 param_2,f32 param_3,u32 param_4,  u64 param_5,long param_6,u64 param_7,s32 param_8,u64 param_9,  int param_10,int param_11);
+u64  FUN_003a56f0(u32 param_1,s32 param_2,f32 param_3,u32 param_4,  s32 param_5,s32 param_6,s32 param_7,s32 param_8,s32 param_9,  int param_10,int param_11);
 u32 FUN_003a5940(int param_1,int param_2);
 void FUN_003a5980(u64 param_1);
 void FUN_003a5ca0(int param_1,int param_2,u32 param_3,int param_4);
@@ -3070,9 +3070,9 @@ FUN_003a5570(u32 param_1,u64 param_2,u64 param_3,u32 param_4,
 
 u64
 
-FUN_003a56f0(u32 param_1,u64 param_2,f32 param_3,u32 param_4,
+FUN_003a56f0(u32 param_1,s32 param_2,f32 param_3,u32 param_4,
 
-            u64 param_5,long param_6,u64 param_7,s32 param_8,u64 param_9,
+            s32 param_5,s32 param_6,s32 param_7,s32 param_8,s32 param_9,
 
             int param_10,int param_11)
 
@@ -3082,7 +3082,7 @@ FUN_003a56f0(u32 param_1,u64 param_2,f32 param_3,u32 param_4,
 
   int iVar1;
 
-  u64 uVar2;
+  int uVar2;
 
   u64 uVar3;
